First hacky prototype chm-to-web script.
Processes the TOC from the hhc file in an unpacked chm
and makes wrapper pages embedding the content with a
navigation bar on the left.

Need some way to rewrite links/images/etc to have the
correct path, navigate the parent frame and fix up the
links using "javascript" generated by hhc.